+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Optional `last_phone_number` parameter on `getWaybill()` for courier AWB confirmation (required by JNE)
+
 ## [2.1.0] - 2026-05-22
 
 ### Added